consolidate indexedmap
