+++ src/agdb/collections.rs
@@ -2,7 +2,6 @@ pub mod bit_set;
 pub mod dictionary;
 pub mod dictionary_storage;
 pub mod indexed_map;
-pub mod indexed_map_storage;
 pub mod map_storage;
 pub mod multi_map;
 pub mod multi_map_storage;
